move/#122: 디자인 시스템 컴포넌트 - 리패키징
