+++ core/designsystem/src/main/java/com/suwiki/core/designsystem/component/badge/SuwikiBadge.kt
