+++ feature/lectureevaluation/editor/src/main/java/com/suwiki/feature/lectureevaluation/editor/lectureevaluation/LectureEvaluationEditorScreen.kt
@@ -36,7 +36,7 @@ import com.suwiki.core.designsystem.component.chips.SuwikiContainedChip
 import com.suwiki.core.designsystem.component.container.SuwikiSelectionContainer
 import com.suwiki.core.designsystem.component.loading.LoadingScreen
 import com.suwiki.core.designsystem.component.ratingbar.SuwikiRatingBar
-import com.suwiki.core.designsystem.component.slider.SuwikiSlider
+import com.suwiki.feature.lectureevaluation.editor.lectureevaluation.component.slider.SuwikiSlider
 import com.suwiki.core.designsystem.component.textfield.SuwikiReviewInputBox
 import com.suwiki.core.designsystem.theme.Primary
 import com.suwiki.core.designsystem.theme.SuwikiTheme
