+++ feature/lectureevaluation/editor/src/main/java/com/suwiki/feature/lectureevaluation/editor/lectureevaluation/component/slider/SuwikiSlider.kt
@@ -1,4 +1,4 @@
-package com.suwiki.core.designsystem.component.slider
+package com.suwiki.feature.lectureevaluation.editor.lectureevaluation.component.slider
 
 import androidx.annotation.IntRange
 import androidx.compose.foundation.interaction.MutableInteractionSource
@@ -24,7 +24,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.suwiki.core.designsystem.component.slider.material.Slider
+import com.suwiki.feature.lectureevaluation.editor.lectureevaluation.component.slider.material.Slider
 import com.suwiki.core.designsystem.theme.Primary
 import com.suwiki.core.designsystem.theme.SuwikiTheme
 import kotlin.math.round
