+++ feature/lectureevaluation/editor/src/main/java/com/suwiki/feature/lectureevaluation/editor/lectureevaluation/component/slider/SuwikiSliderThumb.kt
@@ -1,4 +1,4 @@
-package com.suwiki.core.designsystem.component.slider
+package com.suwiki.feature.lectureevaluation.editor.lectureevaluation.component.slider
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
