+++ feature/lectureevaluation/editor/src/main/java/com/suwiki/feature/lectureevaluation/editor/lectureevaluation/component/slider/material/SliderState.kt
@@ -1,4 +1,4 @@
-package com.suwiki.core.designsystem.component.slider.material
+package com.suwiki.feature.lectureevaluation.editor.lectureevaluation.component.slider.material
 
 import androidx.compose.foundation.MutatePriority
 import androidx.compose.foundation.MutatorMutex
