+++ feature/lectureevaluation/viewerreporter/src/main/java/com/suwiki/feature/lectureevaluation/viewerreporter/component/EvaluationAppBar.kt
@@ -1,22 +1,28 @@
-package com.suwiki.core.designsystem.component.appbar
+package com.suwiki.feature.lectureevaluation.viewerreporter.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.wrapContentHeight
+import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Alignment.Companion.CenterVertically
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.suwiki.core.designsystem.component.filter.SuwikiFilterButton
+import com.suwiki.core.designsystem.R
 import com.suwiki.core.designsystem.theme.GrayFB
+import com.suwiki.core.designsystem.theme.Primary
 import com.suwiki.core.designsystem.theme.SuwikiTheme
+import com.suwiki.core.ui.extension.suwikiClickable
 
 @Composable
-fun SuwikiEvaluationAppBar(
+fun EvaluationAppBar(
   modifier: Modifier = Modifier,
   title: String,
   major: String,
@@ -33,7 +39,7 @@ fun SuwikiEvaluationAppBar(
       text = title,
       style = SuwikiTheme.typography.header1,
     )
-    SuwikiFilterButton(
+    FilterButton(
       text = major,
       onClick = onClickMajor,
       modifier = Modifier
@@ -42,11 +48,40 @@ fun SuwikiEvaluationAppBar(
   }
 }
 
+@Composable
+fun FilterButton(
+  modifier: Modifier = Modifier,
+  text: String,
+  onClick: () -> Unit = {},
+) {
+  Row(
+    modifier = modifier
+      .suwikiClickable(onClick = onClick)
+      .wrapContentHeight()
+      .padding(vertical = 4.dp, horizontal = 8.dp),
+    verticalAlignment = Alignment.CenterVertically,
+  ) {
+    Text(
+      text = text,
+      style = SuwikiTheme.typography.header6,
+      color = Primary,
+    )
+    Icon(
+      painter = painterResource(id = R.drawable.ic_filter_arrow_down),
+      contentDescription = "",
+      tint = Primary,
+      modifier = Modifier
+        .size(24.dp)
+        .padding(vertical = 9.dp, horizontal = 7.dp),
+    )
+  }
+}
+
 @Preview(showSystemUi = true)
 @Composable
-fun SuwikiEvaluationAppBarPreview() {
+fun EvaluationAppBarPreview() {
   SuwikiTheme {
-    SuwikiEvaluationAppBar(
+    EvaluationAppBar(
       title = "강의평가",
       major = "학과명",
       onClickMajor = { /*TODO*/ },
