+++ feature/lectureevaluation/viewerreporter/src/main/java/com/suwiki/feature/lectureevaluation/viewerreporter/detail/component/ExamEvaluationContainer.kt
@@ -1,4 +1,4 @@
-package com.suwiki.core.designsystem.component.container
+package com.suwiki.feature.lectureevaluation.viewerreporter.detail.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -6,7 +6,9 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
@@ -17,46 +19,85 @@ import com.suwiki.core.designsystem.R
 import com.suwiki.core.designsystem.component.badge.BadgeColor
 import com.suwiki.core.designsystem.component.badge.SuwikiBadge
 import com.suwiki.core.designsystem.component.button.SuwikiContainedSmallButton
-import com.suwiki.core.designsystem.component.ratingbar.SuwikiRatingBar
 import com.suwiki.core.designsystem.theme.Black
+import com.suwiki.core.designsystem.theme.Gray95
 import com.suwiki.core.designsystem.theme.SuwikiTheme
 import com.suwiki.core.designsystem.theme.White
 
 @Composable
-fun SuwikiUserReviewContainer(
+fun ExamEvaluationContainer(
   modifier: Modifier = Modifier,
-  content: String,
+  difficulty: String,
+  examType: String,
   semester: String,
-  rating: Float,
+  examInfo: String,
+  content: String,
   isAuthor: Boolean = false,
-  onClickButton: () -> Unit = {},
+  onClickButton: () -> Unit,
 ) {
   val buttonText = if (isAuthor) stringResource(id = R.string.word_edit) else stringResource(id = R.string.word_report)
+
   Column(
-    modifier = modifier
+    modifier = Modifier
       .fillMaxWidth()
       .background(White)
       .padding(24.dp),
-    verticalArrangement = Arrangement.spacedBy(14.dp),
   ) {
     Row(
       modifier = Modifier.fillMaxWidth(),
-      horizontalArrangement = Arrangement.spacedBy(8.dp),
     ) {
       if (isAuthor) {
         SuwikiBadge(
           color = BadgeColor.Blue,
           text = stringResource(id = R.string.word_my),
         )
+        Spacer(modifier = Modifier.width(8.dp))
       }
       SuwikiBadge(
         color = BadgeColor.Gray,
         text = semester,
       )
+      Spacer(modifier = Modifier.width(6.dp))
+      SuwikiBadge(
+        color = BadgeColor.Gray,
+        text = examType,
+      )
       Spacer(modifier = Modifier.weight(1f))
       SuwikiContainedSmallButton(text = buttonText, onClick = onClickButton)
     }
-    SuwikiRatingBar(rating = rating)
+    Spacer(modifier = Modifier.height(10.dp))
+    Row(
+      modifier = modifier.fillMaxWidth(),
+      horizontalArrangement = Arrangement.spacedBy(6.dp),
+    ) {
+      Text(
+        text = stringResource(id = R.string.word_difficulcy),
+        style = SuwikiTheme.typography.caption2,
+        color = Gray95,
+      )
+      Text(
+        text = difficulty,
+        style = SuwikiTheme.typography.caption1,
+        color = Black,
+      )
+    }
+    Spacer(modifier = Modifier.height(2.dp))
+    Row(
+      modifier = modifier.fillMaxWidth(),
+      horizontalArrangement = Arrangement.spacedBy(6.dp),
+    ) {
+      Text(
+        text = stringResource(id = R.string.word_type_exam),
+        style = SuwikiTheme.typography.caption2,
+        color = Gray95,
+      )
+      Text(
+        text = examInfo,
+        style = SuwikiTheme.typography.caption1,
+        color = Black,
+      )
+    }
+    Spacer(modifier = Modifier.height(10.dp))
     Text(
       text = content,
       style = SuwikiTheme.typography.body7,
@@ -67,20 +108,26 @@ fun SuwikiUserReviewContainer(
 
 @Preview
 @Composable
-fun ReviewContainerPreview() {
+fun ExamEvaluationContainerPreview() {
   SuwikiTheme {
     Column {
-      SuwikiUserReviewContainer(
+      ExamEvaluationContainer(
         isAuthor = false,
-        semester = "2023-1",
-        rating = 3.0f,
+        difficulty = "어려움",
+        examInfo = "응용,실습,과제,PPT",
         content = "거의 한 학기 팀플하시는데... 팀원 잘 만나면 잘 모르겠네요. 굉장히 오픈 마인드시긴해요.",
+        onClickButton = {},
+        examType = "중간고사",
+        semester = "2023-1",
       )
-      SuwikiUserReviewContainer(
-        rating = 3.0f,
-        semester = "2023-1",
+      ExamEvaluationContainer(
         isAuthor = true,
+        difficulty = "어려움",
+        examType = "중간고사",
         content = "거의 한 학기 팀플하시는데... 팀원 잘 만나면 잘 모르겠네요. 굉장히 오픈 마인드시긴해요.",
+        onClickButton = {},
+        semester = "2023-1",
+        examInfo = "응용,실습,과제,PPT",
       )
     }
   }
