+++ feature/lectureevaluation/viewerreporter/src/main/java/com/suwiki/feature/lectureevaluation/viewerreporter/detail/component/LectureEvaluationContainer.kt
@@ -1,4 +1,4 @@
-package com.suwiki.core.designsystem.component.container
+package com.suwiki.feature.lectureevaluation.viewerreporter.detail.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -6,9 +6,7 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
@@ -19,85 +17,46 @@ import com.suwiki.core.designsystem.R
 import com.suwiki.core.designsystem.component.badge.BadgeColor
 import com.suwiki.core.designsystem.component.badge.SuwikiBadge
 import com.suwiki.core.designsystem.component.button.SuwikiContainedSmallButton
+import com.suwiki.core.designsystem.component.ratingbar.SuwikiRatingBar
 import com.suwiki.core.designsystem.theme.Black
-import com.suwiki.core.designsystem.theme.Gray95
 import com.suwiki.core.designsystem.theme.SuwikiTheme
 import com.suwiki.core.designsystem.theme.White
 
 @Composable
-fun SuwikiExamReviewContainer(
+fun LectureEvaluationContainer(
   modifier: Modifier = Modifier,
-  difficulty: String,
-  examType: String,
-  semester: String,
-  examInfo: String,
   content: String,
+  semester: String,
+  rating: Float,
   isAuthor: Boolean = false,
-  onClickButton: () -> Unit,
+  onClickButton: () -> Unit = {},
 ) {
   val buttonText = if (isAuthor) stringResource(id = R.string.word_edit) else stringResource(id = R.string.word_report)
-
   Column(
-    modifier = Modifier
+    modifier = modifier
       .fillMaxWidth()
       .background(White)
       .padding(24.dp),
+    verticalArrangement = Arrangement.spacedBy(14.dp),
   ) {
     Row(
       modifier = Modifier.fillMaxWidth(),
+      horizontalArrangement = Arrangement.spacedBy(8.dp),
     ) {
       if (isAuthor) {
         SuwikiBadge(
           color = BadgeColor.Blue,
           text = stringResource(id = R.string.word_my),
         )
-        Spacer(modifier = Modifier.width(8.dp))
       }
       SuwikiBadge(
         color = BadgeColor.Gray,
         text = semester,
       )
-      Spacer(modifier = Modifier.width(6.dp))
-      SuwikiBadge(
-        color = BadgeColor.Gray,
-        text = examType,
-      )
       Spacer(modifier = Modifier.weight(1f))
       SuwikiContainedSmallButton(text = buttonText, onClick = onClickButton)
     }
-    Spacer(modifier = Modifier.height(10.dp))
-    Row(
-      modifier = modifier.fillMaxWidth(),
-      horizontalArrangement = Arrangement.spacedBy(6.dp),
-    ) {
-      Text(
-        text = stringResource(id = R.string.word_difficulcy),
-        style = SuwikiTheme.typography.caption2,
-        color = Gray95,
-      )
-      Text(
-        text = difficulty,
-        style = SuwikiTheme.typography.caption1,
-        color = Black,
-      )
-    }
-    Spacer(modifier = Modifier.height(2.dp))
-    Row(
-      modifier = modifier.fillMaxWidth(),
-      horizontalArrangement = Arrangement.spacedBy(6.dp),
-    ) {
-      Text(
-        text = stringResource(id = R.string.word_type_exam),
-        style = SuwikiTheme.typography.caption2,
-        color = Gray95,
-      )
-      Text(
-        text = examInfo,
-        style = SuwikiTheme.typography.caption1,
-        color = Black,
-      )
-    }
-    Spacer(modifier = Modifier.height(10.dp))
+    SuwikiRatingBar(rating = rating)
     Text(
       text = content,
       style = SuwikiTheme.typography.body7,
@@ -108,26 +67,20 @@ fun SuwikiExamReviewContainer(
 
 @Preview
 @Composable
-fun SuwikiExamReviewContainerPreview() {
+fun ReviewContainerPreview() {
   SuwikiTheme {
     Column {
-      SuwikiExamReviewContainer(
+      LectureEvaluationContainer(
         isAuthor = false,
-        difficulty = "어려움",
-        examInfo = "응용,실습,과제,PPT",
-        content = "거의 한 학기 팀플하시는데... 팀원 잘 만나면 잘 모르겠네요. 굉장히 오픈 마인드시긴해요.",
-        onClickButton = {},
-        examType = "중간고사",
         semester = "2023-1",
+        rating = 3.0f,
+        content = "거의 한 학기 팀플하시는데... 팀원 잘 만나면 잘 모르겠네요. 굉장히 오픈 마인드시긴해요.",
       )
-      SuwikiExamReviewContainer(
-        isAuthor = true,
-        difficulty = "어려움",
-        examType = "중간고사",
-        content = "거의 한 학기 팀플하시는데... 팀원 잘 만나면 잘 모르겠네요. 굉장히 오픈 마인드시긴해요.",
-        onClickButton = {},
+      LectureEvaluationContainer(
+        rating = 3.0f,
         semester = "2023-1",
-        examInfo = "응용,실습,과제,PPT",
+        isAuthor = true,
+        content = "거의 한 학기 팀플하시는데... 팀원 잘 만나면 잘 모르겠네요. 굉장히 오픈 마인드시긴해요.",
       )
     }
   }
