+++ feature/lectureevaluation/viewerreporter/src/main/java/com/suwiki/feature/lectureevaluation/viewerreporter/detail/component/StatisticsContainer.kt
@@ -15,6 +15,7 @@ import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.foundation.layout.wrapContentWidth
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.LinearProgressIndicator
 import androidx.compose.material3.Text
 import androidx.compose.material3.VerticalDivider
 import androidx.compose.runtime.Composable
@@ -22,30 +23,36 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.StrokeCap
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.suwiki.core.designsystem.R
 import com.suwiki.core.designsystem.component.badge.BadgeColor
 import com.suwiki.core.designsystem.component.badge.SuwikiBadge
-import com.suwiki.core.designsystem.component.card.SuwikiReviewGradeCard
+import com.suwiki.core.designsystem.component.ratingbar.SuwikiRatingBar
+import com.suwiki.core.designsystem.shadow.cardShadow
 import com.suwiki.core.designsystem.theme.Black
 import com.suwiki.core.designsystem.theme.Blue10
 import com.suwiki.core.designsystem.theme.Blue100
 import com.suwiki.core.designsystem.theme.Gray6A
 import com.suwiki.core.designsystem.theme.GrayDA
+import com.suwiki.core.designsystem.theme.GrayF6
 import com.suwiki.core.designsystem.theme.GrayFB
 import com.suwiki.core.designsystem.theme.Green10
 import com.suwiki.core.designsystem.theme.Green100
 import com.suwiki.core.designsystem.theme.Orange10
 import com.suwiki.core.designsystem.theme.Orange100
+import com.suwiki.core.designsystem.theme.Primary
 import com.suwiki.core.designsystem.theme.SuwikiTheme
+import com.suwiki.core.designsystem.theme.White
 import com.suwiki.core.model.enums.GradeLevel
 import com.suwiki.core.model.enums.HomeworkLevel
 import com.suwiki.core.model.enums.TeamLevel
 import com.suwiki.core.model.lectureevaluation.lecture.LectureEvaluationExtraAverage
 import com.suwiki.core.model.lectureevaluation.lecture.LectureInfo
 import com.suwiki.core.ui.extension.toText
+import java.util.Locale
 
 @Composable
 fun SuwikiReviewStatisticsContainer(
@@ -116,11 +123,108 @@ fun SuwikiReviewStatisticsContainer(
       }
     }
     Spacer(modifier = Modifier.height(23.dp))
-    SuwikiReviewGradeCard(
+    ReviewGradeCard(
       rating = data.totalAvg,
       honeyRating = data.honeyAvg,
       learningRating = data.learningAvg,
       satisfactionRating = data.satisfactionAvg,
+    )
+  }
+}
+
+@Composable
+fun ReviewGradeCard(
+  modifier: Modifier = Modifier,
+  rating: Float,
+  honeyRating: Float,
+  learningRating: Float,
+  satisfactionRating: Float,
+) {
+  val reviewCountColor = if (rating > 0) Primary else GrayDA
+  val reviewIndicatorColor = if (rating > 0) Black else GrayDA
+  Row(
+    modifier = modifier
+      .fillMaxWidth()
+      .cardShadow()
+      .clip(RoundedCornerShape(10.dp))
+      .background(White)
+      .padding(16.dp),
+    verticalAlignment = Alignment.CenterVertically,
+    horizontalArrangement = Arrangement.spacedBy(20.dp),
+  ) {
+    Column(
+      verticalArrangement = Arrangement.Center,
+      horizontalAlignment = Alignment.CenterHorizontally,
+    ) {
+      Text(
+        text = "%.1f".format(Locale.US, rating),
+        style = SuwikiTheme.typography.header1,
+        color = reviewCountColor,
+      )
+      SuwikiRatingBar(
+        rating = rating,
+      )
+    }
+    VerticalDivider(
+      color = GrayF6,
+      modifier = Modifier
+        .height(49.dp)
+        .width(1.dp),
+    )
+    Column(
+      verticalArrangement = Arrangement.Center,
+      horizontalAlignment = Alignment.CenterHorizontally,
+    ) {
+      ReviewProgressLine(
+        name = stringResource(id = R.string.honey_quality),
+        rating = honeyRating,
+        textColor = reviewIndicatorColor,
+      )
+      ReviewProgressLine(
+        name = stringResource(id = R.string.learning_quality),
+        rating = learningRating,
+        textColor = reviewIndicatorColor,
+      )
+      ReviewProgressLine(
+        name = stringResource(id = R.string.satisfaction_quality),
+        rating = satisfactionRating,
+        textColor = reviewIndicatorColor,
+      )
+    }
+  }
+}
+
+@Composable
+fun ReviewProgressLine(
+  name: String,
+  rating: Float,
+  textColor: Color,
+) {
+  Row(
+    modifier = Modifier
+      .fillMaxWidth(),
+    horizontalArrangement = Arrangement.SpaceBetween,
+    verticalAlignment = Alignment.CenterVertically,
+  ) {
+    Text(
+      text = name,
+      style = SuwikiTheme.typography.caption5,
+      color = textColor,
+    )
+    LinearProgressIndicator(
+      progress = { rating / 5.0f },
+      modifier = Modifier
+        .weight(1f)
+        .height(6.dp)
+        .padding(horizontal = 10.dp)
+        .clip(RoundedCornerShape(4.dp)),
+      color = Primary,
+      strokeCap = StrokeCap.Round,
+    )
+    Text(
+      text = "%.1f".format(Locale.US, rating),
+      style = SuwikiTheme.typography.caption1,
+      color = textColor,
     )
   }
 }
